ClientCommandProvider split from CodeActionsProvider. NbLanguageClient moved to API package.
