+++ java/java.lsp.server/src/org/netbeans/modules/java/lsp/server/db/DBCommandProvider.java
@@ -19,21 +19,16 @@
 package org.netbeans.modules.java.lsp.server.db;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
-import org.eclipse.lsp4j.CodeAction;
-import org.eclipse.lsp4j.CodeActionParams;
 import org.netbeans.api.db.explorer.ConnectionManager;
 import org.netbeans.api.db.explorer.DatabaseConnection;
-import org.netbeans.modules.java.lsp.server.explorer.TreeItem;
 import org.netbeans.modules.java.lsp.server.explorer.TreeNodeRegistry;
-import org.netbeans.modules.java.lsp.server.protocol.CodeActionsProvider;
-import org.netbeans.modules.java.lsp.server.protocol.NbCodeLanguageClient;
-import org.netbeans.modules.parsing.api.ResultIterator;
+import org.netbeans.api.lsp.server.NbCodeLanguageClient;
+import org.netbeans.spi.lsp.server.ClientCommandProvider;
 import org.openide.nodes.Node;
 import org.openide.util.Lookup;
 import org.openide.util.lookup.ServiceProvider;
@@ -42,19 +37,14 @@ import org.openide.util.lookup.ServiceProvider;
  *
  * @author sdedic
  */
-@ServiceProvider(service = CodeActionsProvider.class)
-public class DBCommandProvider extends CodeActionsProvider {
+@ServiceProvider(service = ClientCommandProvider.class)
+public class DBCommandProvider implements ClientCommandProvider {
     private static final String  COMMAND_GET_PREFERRED_CONNECTION = "java.db.preferred.connection";
     
     private static final Set<String> COMMANDS = new HashSet<>(Arrays.asList(
         COMMAND_GET_PREFERRED_CONNECTION
     ));
     
-    @Override
-    public List<CodeAction> getCodeActions(ResultIterator resultIterator, CodeActionParams params) throws Exception {
-        return Collections.emptyList();
-    }
-
     @Override
     public CompletableFuture<Object> processCommand(NbCodeLanguageClient client, String command, List<Object> arguments) {
         if (!COMMAND_GET_PREFERRED_CONNECTION.equals(command)) {
