+++ java/java.lsp.server/src/org/netbeans/modules/java/lsp/server/debugging/attach/AttachNativeConfigurations.java
@@ -26,7 +26,7 @@ import org.eclipse.lsp4j.jsonrpc.messages.ResponseErrorCode;
 
 import org.netbeans.modules.java.lsp.server.debugging.attach.Processes.ProcessInfo;
 import org.netbeans.modules.java.lsp.server.debugging.utils.ErrorUtilities;
-import org.netbeans.modules.java.lsp.server.protocol.NbCodeLanguageClient;
+import org.netbeans.api.lsp.server.NbCodeLanguageClient;
 import org.netbeans.modules.java.lsp.server.protocol.QuickPickItem;
 import org.netbeans.modules.java.lsp.server.protocol.ShowQuickPickParams;
