+++ java/java.lsp.server/src/org/netbeans/modules/java/lsp/server/progress/OperationContext.java
@@ -36,7 +36,7 @@ import java.util.regex.Pattern;
 import org.eclipse.lsp4j.WorkDoneProgressCreateParams;
 import org.eclipse.lsp4j.jsonrpc.messages.Either;
 import org.netbeans.api.progress.ProgressHandle;
-import org.netbeans.modules.java.lsp.server.protocol.NbCodeLanguageClient;
+import org.netbeans.api.lsp.server.NbCodeLanguageClient;
 import org.netbeans.modules.progress.spi.Controller;
 import org.netbeans.modules.progress.spi.InternalHandle;
 import org.netbeans.modules.progress.spi.TaskModel;
